refactor(program-recommend): ProgramRecommendEditor 서브컴포넌트 추출
- ProgramSelectModal 모달 컴포넌트 분리 (modal/)
- ProgramRecommendItem 아이템 카드 분리 (ui/)
- isSameProgram, toRecommendItem 헬퍼로 비교/변환 로직 정리
- ProgramRecommendEditor를 조합 컴포넌트로 축소 (317줄 → 93줄)
